+++ lib/src/endpoints/sources.dart
@@ -17,8 +17,23 @@ class Sources extends Endpoint {
     this.country,
   });
 
+  /// Find sources that display news of this category.
+  /// Possible options are contained in the [Categories]
+  /// enum.
+  /// Defaults to all categories.
   final Categories category;
+
+  /// Find sources that display news in a specific
+  /// language.
+  /// Possible options are: 'ar', 'de', 'en', 'es', 'fr',
+  /// 'he', 'it', 'nl', 'no', 'pt', 'ru', 'se', 'ud', 'zh'.
+  /// Defaults to all languages.
   final String language;
+
+  /// Find sources that display news in a specific country.
+  /// Possible options are contained in the [Countries]
+  /// class.
+  /// Defaults to all countries.
   final String country;
 
   String get url => 'sources';